+++ Changes.md
@@ -9,6 +9,11 @@ Breaking Changes
 1.2.x.x (relative to 1.2.1.0)
 =======
 
+Fixes
+-----
+
+- ImageReader : Made error message more descriptive when trying to access a channel that doesn't exist.
+
 1.2.1.0 (relative to 1.2.0.2)
 =======
 
@@ -270,6 +275,12 @@ Build
 
 1.1.9.x (relative to 1.1.9.2)
 =======
+
+Fixes
+-----
+
+- ImageReader : Made error message more descriptive when trying to access a channel that doesn't exist.
+
 
 1.1.9.2 (relative to 1.1.9.1)
 =======