+++ Changes.md
@@ -1,12 +1,18 @@
 1.x.x.x (relative to 1.2.x.x)
 =======
 
+API
+---
+
+- ViewportGadget : A post-process shader can now be applied to any layer, not just the main one.
+
 Breaking Changes
 ----------------
 
 - GLWidget :
   - A GL context is no longer available in `_resize()`.
   - Removed `BufferOptions.Double`.
+- ViewportGadget : Changed function signature for `setPostProcessShader()` and `getPostProcessShader()`.
 
 1.2.x.x (relative to 1.2.1.0)
 =======